+++ wp-content/plugins/interactieve-map/admin/views/im_admin_checkpoint_edit.php
@@ -60,7 +60,7 @@ $imageFileName = ($_FILES['image']['name']);
 // Count # of uploaded files in array
 $total = count($_FILES['image']['name']);
 
-$emptyFileError = '';
+$fileExists = false;
 
 // Loop through each file
 if ($total > 0) {
@@ -80,22 +80,23 @@ if ($total > 0) {
 
             // Check if filename already exists
             if(file_exists($newFilePath)) {
-                echo '<script>alert("Een van de bestanden word al gebruikt of bestaan al.");</script>';
-                echo '<script>window.history.back();</script>';
-                exit;
+                $fileExists = true;
+                $errors[] = "Een of meerdere gekozen afbeeldingen bestaan al." . '<br>';
+            }
+
+            // Check if filename already exists
+            if(file_exists($uploadPath) && (!empty($fileName))) {
+                $fileExists = true;
             }
 
             if (! in_array($imageFileExtension,$fileExtensions)) {
-                $imageErrors[] = "Dit bestand type is niet mogelijk. Upload een JPEG, JPG of PNG." . '<br>';
+                $errors[] = "Het bestand type " . $imageFileExtension . " is niet mogelijk. Upload een JPEG, JPG of PNG." . '<br>';
             }
 
             // If no errors are found
-            if(empty($imageErrors)) {
+            if (empty($errors) && $fileExists == false) {
                 //Upload the file
                 $didImageUpload = move_uploaded_file($tmpFilePath, $newFilePath);
-            } else {
-                $imageErrors = "Kon bestand niet uploaden, probeer het opnieuw. ";
-                echo ($imageErrors);
             }
         }
     }
@@ -103,33 +104,33 @@ if ($total > 0) {
 
 // If submit
 if (isset($input_array['submit']) && !empty($input_array['submit'])) {
-    // If no file is uploaded, don't active checks for uploaded file
-    if (empty($fileName) && (empty($imageErrors))) {
-        // Refer to different update if no file has been uploaded
-        $checkpoints->update($input_array, $fileName, $imageFileName);
+
+    // If no image is uploaded, don't activate checks for uploaded image
+    if (empty($fileName) && $fileExists == false) {
+        // Refer to different update if no image has been uploaded
+        $checkpoints->update($input_array, $fileName, $imageFileName, $id);
         echo '<script>location.href="?page=interactieve-map-admin";</script>';
         exit;
     }
     else {
         // Check if filename already exists
-        if(file_exists($uploadPath)) {
-            echo '<script>alert("Dit bestaand word al gebruikt of het bestand bestaat al.");</script>';
-            echo '<script>window.history.back();</script>';
-            exit;
+        if(file_exists($uploadPath) && !empty($fileName)) {
+            $fileExists = true;
+            $errors[] = "De icon " . $fileName . ", bestaat al." . '<br>';
         }
 
         // If uploaded file doesn't match the available extensions
-        if (!in_array($fileExtension, $fileExtensions)) {
+        if (!in_array($fileExtension, $fileExtensions) && !empty($fileName)) {
             $errors[] = "Dit bestand type is niet mogelijk. Upload een JPEG, JPG of PNG.";
         }
 
         // If no errors are found
-        if (empty($errors) &&  empty($imageErrors)) {
+        if (empty($errors) &&  empty($imageErrors) && $fileExists === false) {
             $didUpload = move_uploaded_file($fileTmpName, $uploadPath);
 
             // If file has been uploaded, start create function and redirect to overview page after that
             if ($didUpload) {
-                $checkpoints->update($input_array, $fileName, $imageFileName);
+                $checkpoints->update($input_array, $fileName, $imageFileName, $id);
                 echo '<script>location.href="?page=interactieve-map-admin";</script>';
                 exit;
             } else {
@@ -429,8 +430,7 @@ if (isset($_POST['delete']) && !empty($_POST['delete'])) {
         //Set current marker
         marker = new google.maps.Marker({
             position: getLocation,
-            map: map,
-            draggable: true //make it draggable
+            map: map
         });
 
         //Listen for any clicks on the map.
@@ -442,12 +442,7 @@ if (isset($_POST['delete']) && !empty($_POST['delete'])) {
                 //Create the marker.
                 marker = new google.maps.Marker({
                     position: clickedLocation,
-                    map: map,
-                    draggable: true //make it draggable
-                });
-                //Listen for drag events!
-                google.maps.event.addListener(marker, 'dragend', function(event){
-                    markerLocation();
+                    map: map
                 });
             } else{
                 //Marker has already been added, so just change its location.
